Implement EvolutionNodeContext
