+++ src/components/pokemon/evolution-chain/EvolutionTrigger.tsx
@@ -1,7 +1,9 @@
-interface Props {
-  trigger: number;
-}
+import { useEvolutionNode } from './evolution-node.context';
 
-export function EvolutionTrigger({ trigger }: Props) {
+export function EvolutionTrigger() {
+  const { details } = useEvolutionNode();
+
+  const trigger = details?.[0].minLevel || null;
+
   return <div className="z-10 rounded-md bg-gray-100 px-2 py-1">Level {trigger}</div>;
 }
